user check in crud

+++ routers/internshipRouter.js
@@ -18,8 +18,15 @@ router
     isLoggedIn,
     checkRole("company"),
     catchAsync(async (req, res) => {
+      const company = await prisma.company.findUnique({
+        where: {
+          companyID: req.user.profile.emails[0].value,
+        },
+      });
+
       const newListing = await prisma.internship_listing.create({
         data: {
+          companyID: company.companyID,
           ...req.body,
           ...(req.body.startDate
             ? { startDate: new Date(req.body.startDate) }
@@ -48,9 +55,26 @@ router
     isLoggedIn,
     checkRole("company"),
     catchAsync(async (req, res) => {
+      const company = await prisma.company.findUnique({
+        where: {
+          contactEmail: req.user.profile.emails[0].value,
+        },
+      });
+
+      const listing = await prisma.internship_listing.findUnique({
+        where: {
+          companyID: Number(company.companyID),
+        },
+      });
+
+      if (!listing) {
+        return res.status(401).json({ error: "Unauthorized" });
+      }
+
       const updatedListing = await prisma.internship_listing.update({
         where: {
           listingID: Number(req.params.id),
+          companyID: Number(company.companyID),
         },
         data: {
           ...req.body,
@@ -68,9 +92,26 @@ router
     isLoggedIn,
     checkRole("company"),
     catchAsync(async (req, res) => {
+      const company = await prisma.company.findUnique({
+        where: {
+          contactEmail: req.user.profile.emails[0].value,
+        },
+      });
+
+      const listing = await prisma.internship_listing.findUnique({
+        where: {
+          companyID: Number(company.companyID),
+        },
+      });
+
+      if (!listing) {
+        return res.status(401).json({ error: "Unauthorized" });
+      }
+
       const deletedListing = await prisma.internship_listing.delete({
         where: {
           listingID: Number(req.params.id),
+          companyID: Number(company.companyID),
         },
       });
